Hypothesis testing for Ctrl and MB6 populations
Personally, I don't think this comparison is meaningful in any way, since each site being testing was found to be differentially accessible, pair-wise, between Ctrl and MB6.
Testing the entire set of sites is meaningless, but the reviewers and collaborators are asking for it.
It's not factually incorrect, it's just not a meaningful question to ask